generic_edit lib
